requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
mysqlclient==1.4.4
pytz==2019.2
sqlparse==0.3.0
